+++ internal/io/rest/reconcile_test.go
@@ -37,7 +37,7 @@ func TestReconcileExact(t *testing.T) {
 		{"2", "Bubo bubo", 4, "0eeccd70-eaf2-5c51-ad8b-46cfb3db1645", 1, true},
 		{"3", "Pomatomus", 3, "82110143-0b8d-50f6-b34d-e2ae118f4e2e", 1, true},
 		{"4", "Pardosa moesta", 7, "e2fdf10b-6a36-5cc7-b6ca-be4d3b34b21f", 1, true},
-		{"5", "Plantago major var major", 5, "bdfc5d4c-478b-5b3f-8f03-375e4daadc04", 1, true},
+		{"5", "Plantago major var major", 7, "bdfc5d4c-478b-5b3f-8f03-375e4daadc04", 1, true},
 		{"6", "Cytospora ribis mitovirus 2", 1, "bd8cc487-9a28-5910-8d98-38d2b43d1dcb", 0.675, false},
 		{"7", "A-shaped rods", 0, "", 0.0, false},
 		{"8", "Alb. alba", 0, "", 0.0, false},